+++ static/js/dev/adminEdit.js
@@ -1,6 +1,7 @@
 	//Box for size checkbox
 	const sizeCheckbox = document.getElementsByName('size_chois');
 	//Product info
+	const productId = document.getElementById('productId');
 	const productComposition = document.getElementById('productComposition');
 	const productSize = document.getElementById('productSize');
 	const productTitle = document.getElementById('productTitle');
@@ -16,6 +17,7 @@
 	//Button
 	const addProductBtn = document.getElementById('addProduct');
 	const editProductBtn = document.getElementById('editProduct');	
+	const previewImage = document.getElementById('previewImage');
 	const compositions = {
 		wool: { ru: "Шерсть", eng: "wool" },
 		polyester: { ru: "Полиэстер", eng: "polyester" },
@@ -29,10 +31,11 @@
 	}
 	const sizeName = [42,44,46,48,50,52,54,56,58,60];
 
-	let errorText = document.getElementById('error');
-	let successText = document.getElementById('success');
-	let productCompositionArr = [];
-	let sizeArr = [];
+	var errorText = document.getElementById('error');
+	var successText = document.getElementById('success');
+
+	var productCompositionArr = [];
+	var sizeArr = [];
 
 	function getSale(price ,percent) {
 		if (percent >= 0 && percent <= 100 && percent[0] != '0') {
@@ -95,7 +98,7 @@
 			fd.append("productPriceAfterSale", productPriceAfterSale.value);
 			fd.append("productIsAvailability", translateBoolToInt(productIsAvailability.checked));
 			xhr.send(fd);
-			xhr.onreadystatechange = function() { // (3)
+			xhr.onreadystatechange = () => { 
 				if (xhr.readyState != 4) return;
 				if (xhr.status != 200) {
 					console.log(`${xhr.status} : ${xhr.statusText}`);
@@ -113,21 +116,59 @@
 
 	function ajaxEdit(e) {
 		e.preventDefault();
-		console.log(e)
+		if (productTitle.value.length >= 1 && productArticle.value.length >= 1 && productSize.value.length >= 1 && productComposition.value.length >= 1 && productPercent.value.length >= 1 && productPrice.value.length >= 1 && productPriceAfterSale.value.length >= 1) {
+			let xhr = new XMLHttpRequest();
+			let url = '../editProduct';
+			let file = document.getElementById('uploadimage').files[0];
+			let fd = new FormData();
+		
+			xhr.open('POST', url, true);
+			fd.append("uploadimage", file);
+			fd.append("productId", productId.value);
+			fd.append("productTitle", productTitle.value);
+			fd.append("productArticle", productArticle.value);
+			fd.append("productCategory", getSelected(productCategory));
+			fd.append("productBrand", getSelected(productBrand));
+			fd.append("productColour", getSelected(productColour));
+			fd.append("productSeason", getSelected(productSeason));
+			fd.append("productSize", productSize.value);
+			fd.append("productComposition", productComposition.value);
+			fd.append("productIsSale", translateBoolToInt(productIsSale.checked));
+			fd.append("productPercent", productPercent.value);
+			fd.append("productPrice", productPrice.value);
+			fd.append("productPriceAfterSale", productPriceAfterSale.value);
+			fd.append("productIsAvailability", translateBoolToInt(productIsAvailability.checked));
+			xhr.send(fd);
+			xhr.onreadystatechange = () => { // (3)
+				if (xhr.readyState != 4) return;
+				if (xhr.status != 200) {
+					console.log(`${xhr.status} : ${xhr.statusText}`);
+					errorText.textContent = 'Ошибка';
+				} else {
+					if (xhr.responseText == 1) {
+						window.location = '../';
+					}
+				}
+			}			
+		} else {
+			errorText.textContent = "Заполните все поля!";
+		}
 	}
 	function createCheckBox(sizeName) {
 		const appendTo = document.getElementById('size_chois');
-		for (var key of sizeName) {
-			let checkBox = document.createElement('input');
-			let label = document.createElement('label');
-			checkBox.type = 'checkbox';
-			checkBox.value = key;
-			checkBox.name = "size_chois";
-			checkBox.id = `size_${key}`;
-			label.htmlFor = `size_${key}`;
-			label.textContent =  key;
-			appendTo.appendChild(label);
-			appendTo.appendChild(checkBox);
+		if (appendTo) {
+			for (var key of sizeName) {
+				let checkBox = document.createElement('input');
+				let label = document.createElement('label');
+				checkBox.type = 'checkbox';
+				checkBox.value = key;
+				checkBox.name = "size_chois";
+				checkBox.id = `size_${key}`;
+				label.htmlFor = `size_${key}`;
+				label.textContent =  key;
+				appendTo.appendChild(label);
+				appendTo.appendChild(checkBox);
+			}			
 		}
 	}
 
@@ -136,8 +177,8 @@
 	function init() {
 		createCheckBox(sizeName);
 		for (let size of sizeCheckbox) {
-			size.addEventListener('change', function(e){
-				function addToProductSize(arr){
+			size.addEventListener('change', e => {
+				let addToProductSize = arr => {
 					if (arr.length >= 1) {
 						arr.sort(function(a, b) {
 							if (a > b) {
@@ -198,18 +239,21 @@
 		}
 		init() {
 			const div = document.createElement('div');
-			div.className = `composition_box`;
-			this.container.appendChild(div);
-			this.createInput(div, 'checkbox');
-			this.createInput(div, 'text');
-			this.createInput(div, 'button', "Добавить", function(e) { 
-				let id = e.target.id.split('_')[1];
-				window[`composition_${id}`].add(); 
-			});
-			this.createInput(div, 'button', "Удалить", function(e) { 
-				let id = e.target.id.split('_')[1];
-				window[`composition_${id}`].delete(); 
-			});
+			if (this.container) {
+				div.className = `composition_box`;
+				this.container.appendChild(div);
+				this.createInput(div, 'checkbox');
+				this.createInput(div, 'text');
+				this.createInput(div, 'button', "Добавить", e => { 
+					let id = e.target.id.split('_')[1];
+					window[`composition_${id}`].add(); 
+				});
+				this.createInput(div, 'button', "Удалить", e => { 
+					let id = e.target.id.split('_')[1];
+					window[`composition_${id}`].delete(); 
+				});				
+			}
+
 		}
 		renderValue(arr) {
 			productComposition.value = arr.join(", ");
